# compiler(amd64): eliminate a bounds check in memory copy